backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.5
langchain==0.0.184
openai==0.27.0
chromadb==0.3.21
pydantic==1.10.2
python-dotenv==0.19.2
PyPDF2==3.0.1
tiktoken==0.8.0